Unity Project 1
This is the first prototype that I worked over winter break of 12/12/2024 - 1/11/2025.

In this project I learned the basics of Unity's engine: Creating Scripts, Prefabs, Managers, Using GameObjects.

The game I want to create is a turn-based game where you prepare your unit's actions and allow them to perform all of their actions at the same time once you end your turn. The goal of this game is to hopefully innovate on auto-chess battlers and RPG games as both aspects of these genres have heavy inspiration for this game.

This project has multiple issues:
- Uncommented code
- Public Static Manager Classes
- Hard to follow code

This project taught me the importance of keeping code commented and organized. Half of the time I would have multiple tabs open and use my brain to figure out where information was being passed through. For example, to display the unit's abilities when that button is pressed in the character window. When the ability button is pressed, Button Manager would run a method in Unit Manager (which had the information for selected unit) that would run another method in Menu Manager. Menu Manager would call that selected member's method to instantiate ability buttons. This could have easily been simplified if I used UnitManager.Instance.SelectedMember, which I was turned off by because of the amount of times I used public variables.

Other things I learned:
I learned how to zoom in onto unit sprites and successfully displayed a range on a grid for movement and attacking. Those are pretty big achievements, because that can be easily translated to Prototype 3.